+++ node/slack-command-bot/README.md
@@ -24,23 +24,6 @@ Sample `200` Response:
 Hello, World!
 ```
 
-### GET, PUT, PATCH, DELETE /
-
-- Returns a "Learn More" JSON response.
-
-**Response**
-
-Sample `200` Response:
-
-```json
-{
-  "motto": "Build Fast. Scale Big. All in One Place.",
-  "learn": "https://appwrite.io/docs",
-  "connect": "https://appwrite.io/discord",
-  "getInspired": "https://builtwith.appwrite.io"
-}
-```
-
 ## ⚙️ Configuration
 
 | Setting           | Value         |
